+++ crates/kas-core/src/event/cx/cx_pub.rs
@@ -40,28 +40,6 @@ impl EventState {
         self.accesskit_is_enabled
     }
 
-    /// True when access key labels should be shown
-    ///
-    /// (True when Alt is held and no widget has character focus.)
-    ///
-    /// This is a fast check.
-    #[inline]
-    pub fn show_access_labels(&self) -> bool {
-        self.modifiers.alt_key()
-    }
-
-    /// Get whether this widget has `(key_focus, sel_focus)`
-    ///
-    /// -   `key_focus`: implies this widget receives keyboard input
-    /// -   `sel_focus`: implies this widget is allowed to select things
-    ///
-    /// Note that `key_focus` implies `sel_focus`.
-    #[inline]
-    pub fn has_key_focus(&self, w_id: &Id) -> (bool, bool) {
-        let sel_focus = *w_id == self.sel_focus;
-        (sel_focus && self.key_focus, sel_focus)
-    }
-
     /// Check whether a widget is disabled
     ///
     /// A widget is disabled if any ancestor is.
@@ -74,12 +52,6 @@ impl EventState {
             }
         }
         false
-    }
-
-    /// Get the current modifier state
-    #[inline]
-    pub fn modifiers(&self) -> ModifiersState {
-        self.modifiers
     }
 
     /// Access event-handling configuration
@@ -276,207 +248,6 @@ impl EventState {
     #[inline]
     pub fn window_action(&mut self, action: Action) {
         self.action |= action;
-    }
-
-    fn access_layer_for_id(&mut self, id: &Id) -> Option<&mut AccessLayer> {
-        let root = &Id::ROOT;
-        for (k, v) in self.access_layers.range_mut(root..=id).rev() {
-            if k.is_ancestor_of(id) {
-                return Some(v);
-            };
-        }
-        debug_assert!(false, "expected ROOT access layer");
-        None
-    }
-
-    /// Add a new access key layer
-    ///
-    /// This method constructs a new "layer" for access keys: any keys
-    /// added via [`EventState::add_access_key`] to a widget which is a descentant
-    /// of (or equal to) `id` will only be active when that layer is active.
-    ///
-    /// This method should only be called by parents of a pop-up: layers over
-    /// the base layer are *only* activated by an open pop-up.
-    ///
-    /// If `alt_bypass` is true, then this layer's access keys will be
-    /// active even without Alt pressed (but only highlighted with Alt pressed).
-    pub fn new_access_layer(&mut self, id: Id, alt_bypass: bool) {
-        self.access_layers.insert(id, (alt_bypass, HashMap::new()));
-    }
-
-    /// Enable `alt_bypass` for layer
-    ///
-    /// This may be called by a child widget during configure to enable or
-    /// disable alt-bypass for the access-key layer containing its access keys.
-    /// This allows access keys to be used as shortcuts without the Alt
-    /// key held. See also [`EventState::new_access_layer`].
-    pub fn enable_alt_bypass(&mut self, id: &Id, alt_bypass: bool) {
-        if let Some(layer) = self.access_layer_for_id(id) {
-            layer.0 = alt_bypass;
-        }
-    }
-
-    /// Register `id` as handler of an access `key`
-    ///
-    /// An *access key* (also known as mnemonic) is a shortcut key able to
-    /// directly open menus, activate buttons, etc. By default, when the user
-    /// presses (and holds) key <kbd>Alt</kbd>, access keys in labels will be
-    /// underlined and when the user presses the corresponding key then the
-    /// widget registered as handler for that access key will receive navigation
-    /// focus and [`Command::Activate`].
-    ///
-    /// If `id` itself cannot support navigation focus or handle
-    /// [`Command::Activate`], then ancestors of `id` will have the opportunity
-    /// to receive navigation focus and handle this [`Event::Command`].
-    ///
-    /// If multiple widgets attempt to register themselves as handlers of the
-    /// same `key`, then only the first succeeds.
-    ///
-    /// [`Self::enable_alt_bypass`] allows usage of access keys without
-    /// <kbd>Alt</kbd>.
-    ///
-    /// Note that access keys may be automatically derived from labels:
-    /// see [`crate::text::AccessString`].
-    ///
-    /// Access keys are added to the layer with the longest path which is
-    /// an ancestor of `id`. This usually means that if the widget is part of a
-    /// pop-up, the key is only active when that pop-up is open.
-    /// See [`EventState::new_access_layer`].
-    ///
-    /// This should only be called from [`Events::configure`].
-    #[inline]
-    pub fn add_access_key(&mut self, id: &Id, key: Key) {
-        if let Some(layer) = self.access_layer_for_id(id) {
-            layer.1.entry(key).or_insert_with(|| id.clone());
-        }
-    }
-
-    /// Visually depress a widget via a key code
-    ///
-    /// When a button-like widget is activated by a key it may call this to
-    /// ensure the widget is visually depressed until the key is released.
-    /// The widget will not receive a notification of key-release but will be
-    /// redrawn automatically.
-    ///
-    /// Note that keyboard shortcuts and mnemonics should usually match against
-    /// the "logical key". [`PhysicalKey`] is used here since the the logical key
-    /// may be changed by modifier keys.
-    ///
-    /// Does nothing when `code` is `None`.
-    pub fn depress_with_key(&mut self, id: Id, code: impl Into<Option<PhysicalKey>>) {
-        fn inner(state: &mut EventState, id: Id, code: PhysicalKey) {
-            if state
-                .key_depress
-                .get(&code)
-                .map(|target| *target == id)
-                .unwrap_or(false)
-            {
-                return;
-            }
-
-            state.key_depress.insert(code, id.clone());
-            state.action(id, Action::REDRAW);
-        }
-
-        if let Some(code) = code.into() {
-            inner(self, id, code);
-        }
-    }
-
-    /// Request keyboard input focus
-    ///
-    /// When granted, the widget will receive [`Event::KeyFocus`] followed by
-    /// [`Event::Key`] for each key press / release. Note that this disables
-    /// translation of key events to [`Event::Command`] while key focus is
-    /// active.
-    ///
-    /// Providing an [`ImePurpose`] enables Input Method Editor events
-    /// (see [`Event::ImeFocus`]). TODO: this feature is incomplete; winit does
-    /// not currently support setting surrounding text.
-    ///
-    /// The `source` parameter is used by [`Event::SelFocus`].
-    ///
-    /// Key focus implies sel focus (see [`Self::request_sel_focus`]) and
-    /// navigation focus.
-    #[inline]
-    pub fn request_key_focus(&mut self, target: Id, ime: Option<ImePurpose>, source: FocusSource) {
-        if self.nav_focus.as_ref() != Some(&target) {
-            self.set_nav_focus(target.clone(), source);
-        }
-
-        self.pending_sel_focus = Some(PendingSelFocus {
-            target: Some(target),
-            key_focus: true,
-            ime,
-            source,
-        });
-    }
-
-    /// End Input Method Editor focus on `target`, if present
-    #[inline]
-    pub fn cancel_ime_focus(&mut self, target: Id) {
-        if let Some(pending) = self.pending_sel_focus.as_mut() {
-            if pending.target.as_ref() == Some(&target) {
-                pending.ime = None;
-            }
-        } else if self.ime.is_some() && self.sel_focus.as_ref() == Some(&target) {
-            self.pending_sel_focus = Some(PendingSelFocus {
-                target: Some(target),
-                key_focus: self.key_focus,
-                ime: None,
-                source: FocusSource::Synthetic,
-            });
-        }
-    }
-
-    /// Set IME cursor area
-    ///
-    /// This should be called after receiving [`Event::ImeFocus`], and any time
-    /// that the `rect` parameter changes, until [`Event::LostImeFocus`] is
-    /// received.
-    ///
-    /// This sets the text cursor's area, `rect`, relative to the widget's own
-    /// coordinate space. If never called, then the widget's whole rect is used.
-    ///
-    /// This does nothing if `target` does not have IME-enabled input focus.
-    #[inline]
-    pub fn set_ime_cursor_area(&mut self, target: &Id, rect: Rect) {
-        if self.ime.is_some() && self.sel_focus.as_ref() == Some(target) {
-            self.ime_cursor_area = rect;
-        }
-    }
-
-    /// Request selection focus
-    ///
-    /// To prevent multiple simultaneous selections (e.g. of text) in the UI,
-    /// only widgets with "selection focus" are allowed to select things.
-    /// Selection focus is implied by character focus. [`Event::LostSelFocus`]
-    /// is sent when selection focus is lost; in this case any existing
-    /// selection should be cleared.
-    ///
-    /// The `source` parameter is used by [`Event::SelFocus`].
-    ///
-    /// Selection focus implies navigation focus.
-    ///
-    /// When key focus is lost, [`Event::LostSelFocus`] is sent.
-    #[inline]
-    pub fn request_sel_focus(&mut self, target: Id, source: FocusSource) {
-        if self.nav_focus.as_ref() != Some(&target) {
-            self.set_nav_focus(target.clone(), source);
-        }
-
-        if let Some(ref pending) = self.pending_sel_focus
-            && pending.target.as_ref() == Some(&target)
-        {
-            return;
-        }
-
-        self.pending_sel_focus = Some(PendingSelFocus {
-            target: Some(target),
-            key_focus: false,
-            ime: None,
-            source,
-        });
     }
 
     /// Send a message to `id`
